feat: tag index pages, shared posts loader, carousel link fixes
- Add /tags/[tag] with server-rendered list and nodejs runtime; layout matches blog chrome.
- Extract getAllPostsMeta() for GET /api/posts and tag filtering (DRY).
- Replace CSS mask carousels with CarouselEdgeFog gradients so tag/post links stay clickable.
- Skip marquee touch handlers when the target is inside a[href] (mobile taps on tags).

Made-with: Cursor
