+++ app/globals.css
@@ -32,36 +32,6 @@ body {
   scrollbar-width: none;
 }
 
-/* Horizontal carousel / marquee: soft edge fade (“mlha”) on narrow viewports too */
-.mask-fade-x-carousel {
-  -webkit-mask-image: linear-gradient(
-    to right,
-    transparent 0%,
-    rgba(0, 0, 0, 0.08) 6%,
-    rgba(0, 0, 0, 0.35) 14%,
-    #000 24%,
-    #000 76%,
-    rgba(0, 0, 0, 0.35) 86%,
-    rgba(0, 0, 0, 0.08) 94%,
-    transparent 100%
-  );
-  mask-image: linear-gradient(
-    to right,
-    transparent 0%,
-    rgba(0, 0, 0, 0.08) 6%,
-    rgba(0, 0, 0, 0.35) 14%,
-    #000 24%,
-    #000 76%,
-    rgba(0, 0, 0, 0.35) 86%,
-    rgba(0, 0, 0, 0.08) 94%,
-    transparent 100%
-  );
-  -webkit-mask-size: 100% 100%;
-  mask-size: 100% 100%;
-  -webkit-mask-repeat: no-repeat;
-  mask-repeat: no-repeat;
-}
-
 a {
   text-decoration: none;
 }
